requirements.txt
langchain~=0.0.101
torch~=2.0.1+cu117
torchvision~=0.15.2+cu117
accelerate
addict
albumentations
basicsr
controlnet-aux
diffusers>=0.19.3
einops~=0.6.1
gradio~=3.39.0
imageio
imageio-ffmpeg
invisible-watermark
kornia
numpy~=1.24.1
omegaconf
open_clip_torch
openai
opencv-python~=4.8.0.74
prettytable
safetensors
streamlit
test-tube
timm~=0.9.5
torchmetrics
transformers~=4.31.0
webdataset
yapf
gradio

setuptools~=67.8.0
requests~=2.28.1
pillow~=9.3.0
matplotlib~=3.7.2
wget~=3.2
groundingdino~=0.1.0